+++ src/Framework/Bootstrap/Setup/Properties.php
@@ -24,6 +24,7 @@ use Gacela\Framework\Event\Dispatcher\EventDispatcherInterface;
  * @psalm-import-type ServiceAliasMap from ContainerConfigurationInterface
  * @psalm-import-type ServicesToExtendMap from ContainerConfigurationInterface
  * @psalm-import-type HandlerRegistriesMap from ContainerConfigurationInterface
+ * @psalm-import-type PluginStacksMap from ContainerConfigurationInterface
  * @psalm-import-type TagsMap from ContainerConfigurationInterface
  * @psalm-import-type AfterResolvingMap from ContainerConfigurationInterface
  * @psalm-import-type DefinitionSources from ContainerConfigurationInterface
@@ -109,6 +110,9 @@ final class Properties
 
     /** @var ?HandlerRegistriesMap */
     public ?array $handlerRegistries = null;
+
+    /** @var ?PluginStacksMap */
+    public ?array $pluginStacks = null;
 
     /** @var ?TagsMap */
     public ?array $tags = null;